Created all_articles() view
In case of GET, all_articles() can render a JSON of all the existing articles.
In case of POST, the user has to enter data on the webpage and if it is valid, it will create it, if it does not exist already, or update it if it already exists and then saves the changes to the database.
Also, shows the new data along with code. If not valid, it shows the errors along with code.